+++ src/DocumentBuilder/vpg_enum_class_reader.cpp
@@ -255,6 +255,7 @@ std::wstring VPGEnumClassReader::_GetCommand(const std::wstring &cppCode, const 
 void VPGEnumClassReader::_ParseProperties(const std::wstring &cppCode, size_t &pos, std::shared_ptr<VPGEnumClass>enumClass) const
 {
     TRY
+        _EnumValue = -1;
         while (pos < cppCode.size()) {
             std::wstring name = _GetEnum(cppCode, pos);
             Trim(name);
@@ -265,9 +266,24 @@ void VPGEnumClassReader::_ParseProperties(const std::wstring &cppCode, size_t &p
             property->_Enum = name;
             GetNextCharPos(cppCode, pos, false);
             if (cppCode[pos] == L'=') {
-                GetNextCharPos(cppCode, pos, false);
+                std::wstring enumValueStr = L"";
+                for (; pos < cppCode.size(); pos++) {
+                    if (std::iswdigit(cppCode[pos]))
+                        enumValueStr += cppCode[pos];
+                    else {
+                        if (!enumValueStr.empty())
+                            break;
+                    }
+                }
+                if (!enumValueStr.empty()) {
+                    _EnumValue = std::stod(enumValueStr);
+                    _EnumValue--;
+                }
+                pos--;
                 GetNextCharPos(cppCode, pos, false);
             }
+            _EnumValue++;
+            property->_EnumValue = _EnumValue;
             if (cppCode[pos] == L',')
                 GetNextCharPos(cppCode, pos, false);
             if (IsStartWith(cppCode, L"//", pos) || IsStartWith(cppCode, L"/*", pos)) {