+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+### [v0.1.4] - 2024-08-25: Java Bridge: Generation - Java Bridege for Enum
+- Generate Java Enum according to C++ Enum Class under Type Workspace.
+
 ### [v0.1.3] - 2024-07-28: Java Bridge: Generation - Dll Interface for Property Accessor
 - Remove support of std::string, all change to use std::wstring
 - Generate Dll Interface for Property Accessor